requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

# Google Generative AI for Gemini
google-generativeai==0.3.2

# Image Processing
Pillow==10.1.0

# HTTP client for testing
aiohttp==3.9.1